requirements.txt
Django
PyMySQL
mysqlclient
Pillow
djangorestframework